added timeout functionality to tgen tranfers
this occurs as a polling event every time the heartbeat event is
run. this is not ideal because: 1) we may wish to change the
the heartbeat log period independent of the timeout check, and
2) an asynchronous approach to checking for timeouts using the
TGenTimeout facility would be more efficient. Also, the timeout
is currently defined to a default of 60 seconds and cannot be
changed without recompiling.

Still, this is a good start.

refs #241