differentiate stalled transfers from hard timeouts
a new 'stallout' option can now be configured in the tgen xml file.
it sets a time when a transfer times out because no progress has
been made. the 'timeout' option now specifies a hard timeout after
which a transfer will timeout if it is not finished no matter its
status.

'stallout' can be used in the same places that 'timeout' could,
i.e., in the `start` action to set a default for all transfers,
or in each `transfer` action to independently configure each.

closes #278