merge thread and process functionality, remove shadow interface funcs
portable threads will be handled at the virtual process layer,
which allowed us to remove a lot of unneccesary code including
the virtual thread object and related code. in our new model,
we will simply call main in the dynamic executable, and then
pth will handle blocking etc. so we no longer need to ask the
plugin to register any functions or be aware that it is running
in shadow in any other way. therefore, we also removed the
interface functions that shadow previously exposed to plugins,
and removed the interface header files defining the api.
finally, we updated tgen plugin to work with this new mode.

refs #236