TGen no longer blocks epoll events on large reads/writes
Rather than reading and writing sockets until they block, we
instead read and write a default amount and then we cycle back
to the kernel. This allows us to check for any new events that
might have arrived in the meantime, and also to round-robin
through our sockets so we don't starve any of them.