Support sending Markov model as path instead of graphml
When the client starts a stream, there is now a new option that
instructs the client to send the configured path to the packet
Markov model rather than sending a string representation of the
model in graphml format. The server must have the model at the
same path or an error will occur.

The new option is 'packetmodelmode' and acceptable values are
'graphml' (the default) or 'path'.

This commit also updates some of the example client tgen configs.