Fix invalid memory access in io timeout check loop
When iterating through the hash table values list, we were
modifying the hash table when we deregister a child that has a
timeout. Such modification invalidates the original child list,
and results in invalid memory access.

Fixed by keeping a list of the descriptors to deregsiter, and
only deregister them after we finish iterating the table.