Stop infinite loops when missing newline
When trying to read a line from the recv buffer, if the current
data did not contain a newline character, then there was a bug
that prevented us from reading more data from the kernel into
the recv buffer in order to hopefully read the newline. This was
causing infinite loops and is now fixed.