Remove dependency on shutdown()
Two tgen endpoints will now compute the number of bytes they
expect to receive from the other end based on the shared Markov
model (that they synchronized with the same seed). Therefore,
each node no longer needs to wait for the FIN that would normally
be send after a shutdown(SHUT_WR) call in order to detect that
a stream is finished.

This was done because Tor will completely close a stream if it
receives a FIN from either the client or server ends. See:
https://lists.torproject.org/pipermail/tor-dev/2019-April/013774.html

We also slightly changed the time tracking logic in order to keep
track of last bytes and first bytes more consistently.