Add info on pending streams and flows to heartbeat message
The driver will now be notified when traffic, flow, and stream
objects are created. This information is printed in the heartbeat
message, along with the number pending objects of each type (those
that are created but not yet completed). We can then estimate the
amount of state stored over time.