Compute unix_ts_start via created-ts and now-ts.
So far we computed stream start time by subtracting the largest number
of elapsed microseconds from the stream end time. This worked well for
successful streams, but not so well for errors where the error might
have occurred long after recording the microsecond timestamp of a
successful substep.

With this change we subtract the difference between created-ts and
now-ts from the stream end time for a more accurate stream start time.
Only if these timestamps are unknown we fall back to setting start and
end time to the same value.

This issue first came up in:

https://gitlab.torproject.org/tpo/metrics/onionperf/-/issues/30362