Replace usage of private igraph_i_attribute_gettype
This fixes the build for igraph 0.9.0+, in which
`igraph_i_attribute_gettype` has been removed:
https://github.com/igraph/igraph/commit/463cc7c1bbddce87f772672bcc2f528b42f0f034

`igraph_cattribute_table.gettype` appears to be the only viable
alternative. It looks like it's intended to be called by igraph itself,
not by users of igraph, but there doesn't appear to be any other API for
checking attribute types.