Fix CI dependencies
When installing both zlib and zlib-devel, dnf thinks that we want the best
possible version of both packages. This is a problem if the best versions of
zlib and zlib-devel are incompatible. Instead, we should install the best
version of zlib-devel, and use whatever zlib version is compatible with that.
Also removed some other packages where we also specify devel versions.