Specify igraph < 0.10.0
igraph 0.10.0 has some API changes that we don't handle. We may be able
to support it through some #ifdef macros, as we already do for 0.9.x,
though the devs suggest this might be painful.

For now, having cmake bail out and explain that the igraph version is
incompatible is better than having a bunch of compiler errors.

See https://github.com/shadow/tgen/issues/44